+++ java/client/src/main/java/glide/managers/CommandManager.java
@@ -2,6 +2,7 @@
 package glide.managers;
 
 import glide.api.models.ClusterTransaction;
+import glide.api.models.Script;
 import glide.api.models.Transaction;
 import glide.api.models.configuration.RequestRoutingConfiguration.ByAddressRoute;
 import glide.api.models.configuration.RequestRoutingConfiguration.Route;
@@ -12,6 +13,7 @@ import glide.api.models.exceptions.ClosingException;
 import glide.api.models.exceptions.RequestException;
 import glide.connectors.handlers.CallbackDispatcher;
 import glide.connectors.handlers.ChannelHandler;
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +23,7 @@ import redis_request.RedisRequestOuterClass.Command.ArgsArray;
 import redis_request.RedisRequestOuterClass.RedisRequest;
 import redis_request.RedisRequestOuterClass.RequestType;
 import redis_request.RedisRequestOuterClass.Routes;
+import redis_request.RedisRequestOuterClass.ScriptInvocation;
 import redis_request.RedisRequestOuterClass.SimpleRoutes;
 import redis_request.RedisRequestOuterClass.SlotTypes;
 import response.ResponseOuterClass.Response;
@@ -82,6 +85,25 @@ public class CommandManager {
             Transaction transaction, RedisExceptionCheckedFunction<Response, T> responseHandler) {
 
         RedisRequest.Builder command = prepareRedisRequest(transaction);
+        return submitCommandToChannel(command, responseHandler);
+    }
+
+    /**
+     * Build a Script (by hash) request to send to Redis.
+     *
+     * @param script Lua script hash object
+     * @param keys The keys that are used in the script
+     * @param args The arguments for the script
+     * @param responseHandler The handler for the response object
+     * @return A result promise of type T
+     */
+    public <T> CompletableFuture<T> submitScript(
+            Script script,
+            List<String> keys,
+            List<String> args,
+            RedisExceptionCheckedFunction<Response, T> responseHandler) {
+
+        RedisRequest.Builder command = prepareRedisRequest(script, keys, args);
         return submitCommandToChannel(command, responseHandler);
     }
 
@@ -164,6 +186,29 @@ public class CommandManager {
 
         RedisRequest.Builder builder =
                 RedisRequest.newBuilder().setTransaction(transaction.getProtobufTransaction().build());
+
+        return builder;
+    }
+
+    /**
+     * Build a protobuf Script Invoke request.
+     *
+     * @param script Redis Script
+     * @param keys keys for the Script
+     * @param args args for the Script
+     * @return An uncompleted request. {@link CallbackDispatcher} is responsible to complete it by
+     *     adding a callback id.
+     */
+    protected RedisRequest.Builder prepareRedisRequest(
+            Script script, List<String> keys, List<String> args) {
+        RedisRequest.Builder builder =
+                RedisRequest.newBuilder()
+                        .setScriptInvocation(
+                                ScriptInvocation.newBuilder()
+                                        .setHash(script.getHash())
+                                        .addAllKeys(keys)
+                                        .addAllArgs(args)
+                                        .build());
 
         return builder;
     }