+++ python/python/tests/test_transaction.py
@@ -31,7 +31,6 @@ async def transaction_test(
     key6 = "{{{}}}:{}".format(keyslot, get_random_string(3))
     key7 = "{{{}}}:{}".format(keyslot, get_random_string(3))
     key8 = "{{{}}}:{}".format(keyslot, get_random_string(3))
-    key9 = "{{{}}}:{}".format(keyslot, get_random_string(3))
 
     value = datetime.now(timezone.utc).strftime("%m/%d/%Y, %H:%M:%S")
     value2 = get_random_string(5)
@@ -48,6 +47,8 @@ async def transaction_test(
     args.append("string")
     transaction.echo(value)
     args.append(value)
+    transaction.strlen(key)
+    args.append(len(value))
 
     transaction.persist(key)
     args.append(False)
@@ -115,12 +116,6 @@ async def transaction_test(
 
     transaction.client_getname()
     args.append(None)
-
-    value9 = "test string for strlen"
-    transaction.set(key9, value9)
-    args.append(OK)
-    transaction.strlen(key9)
-    args.append(len(value9))
 
     transaction.lpush(key5, [value, value, value2, value2])
     args.append(4)